# Improve the performance of logging.