--- a/src/input_output/FGLog.h
+++ b/src/input_output/FGLog.h
@@ -152,20 +152,22 @@ public:
 
   void SetLevel(LogLevel level) override;
   void FileLocation(const std::string& filename, int line) override
-  { out << std::endl << "In file " << filename << ": line" << line << std::endl; }
+  { buffer << std::endl << "In file " << filename << ": line" << line << std::endl; }
   void Format(LogFormat format) override;
   void Flush(void) override {
+    out << buffer.str();
     out.flush();
-    out.clear();
+    buffer.str("");
   }
 
   void Message(const std::string& message) override {
     // if (level < min_level) return;
-    out << message;
+    buffer << message;
   }
 
 private:
   std::ostream out;
+  std::ostringstream buffer;
 };
 } // namespace JSBSim
 #endif