--- a/src/input_output/FGLog.h
+++ b/src/input_output/FGLog.h
@@ -39,7 +39,6 @@ INCLUDES
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*/
 
 #include <iomanip>
-#include <iostream>
 #include <memory>
 #include <sstream>
 #include <string>
@@ -148,17 +147,10 @@ public:
 class JSBSIM_API FGLogConsole : public FGLogger
 {
 public:
-  FGLogConsole() : out(std::cout.rdbuf()) {}
-
-  void SetLevel(LogLevel level) override;
   void FileLocation(const std::string& filename, int line) override
   { buffer << std::endl << "In file " << filename << ": line" << line << std::endl; }
   void Format(LogFormat format) override;
-  void Flush(void) override {
-    out << buffer.str();
-    out.flush();
-    buffer.str("");
-  }
+  void Flush(void) override;
 
   void Message(const std::string& message) override {
     // if (level < min_level) return;
@@ -166,7 +158,6 @@ public:
   }
 
 private:
-  std::ostream out;
   std::ostringstream buffer;
 };
 } // namespace JSBSim