@seanmcleod 2024-08-17
Looks good to merge. I'll try it out on Windows and confirm the performance difference.