@bcoconni 2024-08-18
I've added a commit to remove the now useless member `std::ostream out`. It simplifies the code.