@bcoconni 2024-08-18
> Looks good to merge. I'll try it out on Windows and confirm the performance difference.

Done.