@seanmcleod 2024-08-18
Looks much better.

From https://github.com/JSBSim-Team/jsbsim/pull/1136#issuecomment-2286717524

| Debug Level | Before (s) | After (s) |
| --- | --- | --- |
| 0 |  0.204 | 0.205 |
| 1 |  0.221 | 0.213 |
| 2 | 0.284 | 3.275 |
| 9 | 4.668 | 4.754 |

And now after this PR.

| Debug Level | Before (s) | After (s) |
| --- | --- | --- |
| 0 |  0.247 | 0.246 |
| 1 |  0.267 | 0.261 |
| 2 | 0.300 | 0.433 |
| 9 | 5.042 | 5.134 |

So the ~10x slower performance I saw previously for debug levels 2 and 9 is now gone.